pyproject.toml
[project]
name = "web-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "fastapi[standard]>=0.128.0",
    "psycopg[binary]>=3.3.2",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.45",
]
